Modularize conditional statements
